Version 2
- Pinch gesture developed, now all feasible moves are actually possible.
- Webcam source printed on screen with a grid board drawn on it, to be more usable for the user.
- Moves list added to the board frame.
- Game can be saved by pressing "s".
- Saved game can be loaded by pressing "l".
