+++ game.pgn
@@ -6,4 +6,4 @@
 [Black "?"]
 [Result "*"]
 
-1. e3 e5 2. f4 Nc6 3. fxe5 Nxe5 4. d4 Ng6 5. e4 f6 *
+1. e3 e5 2. Nf3 f5 *
